PrettyTable Output
Script outputs a PrettyTable. Airline Code csv created to match the UniqueCarrier in 2008.csv to the full airline name. Arrival and Departure Board functions made to display the Board. Additional arrival/departure functions made to take an airport as input so the Board functions can use the desired Itinerary information